backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.13.0
django-filter==23.1
Pillow==9.0.0
psycopg2-binary==2.9.3 
gunicorn==20.1.0
djoser==2.1.0
webcolors==1.11.1
PyYAML==6.0